Move attribute models to a new attribute module
