+++ saleor/graphql/csv/tests/mutations/test_export_products.py
@@ -3,9 +3,9 @@ from unittest.mock import ANY, patch
 import graphene
 import pytest
 
+from .....attribute.models import Attribute
 from .....csv import ExportEvents
 from .....csv.models import ExportEvent
-from .....product.models import Attribute
 from .....warehouse.models import Warehouse
 from ....tests.utils import get_graphql_content
 from ...enums import ExportScope, FileTypeEnum, ProductFieldEnum
