+++ saleor/graphql/page/dataloaders.py
@@ -2,9 +2,9 @@ from collections import defaultdict
 
 from promise import Promise
 
+from ...attribute.models import AssignedPageAttribute, AttributePage, AttributeValue
 from ...core.permissions import PagePermissions
 from ...page.models import Page, PageType
-from ...product.models import AssignedPageAttribute, AttributePage, AttributeValue
 from ..core.dataloaders import DataLoader
 from ..product.dataloaders.attributes import (
     AttributesByAttributeId,
