+++ saleor/graphql/product/enums.py
@@ -1,6 +1,6 @@
 import graphene
 
-from ...product import AttributeInputType, AttributeType
+from ...attribute import AttributeInputType, AttributeType
 from ..core.enums import to_enum
 
 AttributeInputTypeEnum = to_enum(AttributeInputType)
