+++ saleor/graphql/product/mutations/products.py
@@ -1,6 +1,6 @@
 import datetime
 from collections import defaultdict
-from typing import Iterable, List, Tuple, Union
+from typing import TYPE_CHECKING, Iterable, List, Tuple, Union
 
 import graphene
 from django.core.exceptions import ObjectDoesNotExist, ValidationError
@@ -10,12 +10,13 @@ from django.utils.text import slugify
 from graphene.types import InputObjectType
 from graphql_relay import from_global_id
 
+from ....attribute import AttributeType
 from ....core.exceptions import PermissionDenied
 from ....core.permissions import ProductPermissions, ProductTypePermissions
 from ....order import OrderStatus, models as order_models
 from ....page import models as page_models
 from ....page.error_codes import PageErrorCode
-from ....product import AttributeType, models
+from ....product import models
 from ....product.error_codes import ProductErrorCode
 from ....product.tasks import (
     update_product_minimal_variant_price_task,
@@ -63,6 +64,9 @@ from ..utils import (
     validate_attributes_input_for_variant,
 )
 from .common import ReorderInput
+
+if TYPE_CHECKING:
+    from ....attribute import models as attribute_models
 
 
 class CategoryInput(graphene.InputObjectType):
@@ -588,7 +592,7 @@ class ProductCreateInput(ProductInput):
     )
 
 
-T_INPUT_MAP = List[Tuple[models.Attribute, List[str]]]
+T_INPUT_MAP = List[Tuple["attribute_models.Attribute", List[str]]]
 T_INSTANCE = Union[models.Product, models.ProductVariant, page_models.Page]
 
 
@@ -619,7 +623,7 @@ class AttributeAssignmentMixin:
     ):
         """Retrieve attributes nodes from given global IDs and/or slugs."""
         qs = qs.filter(Q(pk__in=pks) | Q(slug__in=slugs))
-        nodes = list(qs)  # type: List[models.Attribute]
+        nodes: List["attribute_models.Attribute"] = list(qs)
 
         if not nodes:
             raise ValidationError(
@@ -669,7 +673,9 @@ class AttributeAssignmentMixin:
         return int(internal_id)
 
     @classmethod
-    def _pre_save_values(cls, attribute: models.Attribute, values: List[str]):
+    def _pre_save_values(
+        cls, attribute: "attribute_models.Attribute", values: List[str]
+    ):
         """Lazy-retrieve or create the database objects from the supplied raw values."""
         get_or_create = attribute.values.get_or_create
         return tuple(
