+++ saleor/graphql/product/tests/test_product_sorting_attributes.py
@@ -4,7 +4,8 @@ from decimal import Decimal
 import graphene
 import pytest
 
-from ....product import AttributeType, models as product_models
+from ....attribute import AttributeType, models as attribute_models
+from ....product import models as product_models
 from ....product.utils.attributes import associate_attribute_values_to_instance
 from ...tests.utils import get_graphql_content
 
@@ -49,7 +50,7 @@ def products_structures(category):
 
     assert product_models.Product.objects.count() == 0
 
-    in_multivals = product_models.AttributeInputType.MULTISELECT
+    in_multivals = attribute_models.AttributeInputType.MULTISELECT
 
     pt_apples, pt_oranges, pt_other = list(
         product_models.ProductType.objects.bulk_create(
@@ -68,18 +69,18 @@ def products_structures(category):
     )
 
     colors_attr, trademark_attr, dummy_attr = list(
-        product_models.Attribute.objects.bulk_create(
+        attribute_models.Attribute.objects.bulk_create(
             [
-                product_models.Attribute(
+                attribute_models.Attribute(
                     name="Colors",
                     slug="colors",
                     input_type=in_multivals,
                     type=AttributeType.PRODUCT_TYPE,
                 ),
-                product_models.Attribute(
+                attribute_models.Attribute(
                     name="Trademark", slug="trademark", type=AttributeType.PRODUCT_TYPE
                 ),
-                product_models.Attribute(
+                attribute_models.Attribute(
                     name="Dummy", slug="dummy", type=AttributeType.PRODUCT_TYPE
                 ),
             ]
@@ -504,10 +505,10 @@ def test_sort_product_not_having_attribute_data(api_client, category, count_quer
     )
 
     # Assign an attribute to the product type
-    attribute = product_models.Attribute.objects.create(
+    attribute = attribute_models.Attribute.objects.create(
         name="Kind", slug="kind", type=AttributeType.PRODUCT_TYPE
     )
-    value = product_models.AttributeValue.objects.create(
+    value = attribute_models.AttributeValue.objects.create(
         name="Value", slug="value", attribute=attribute
     )
     product_type.product_attributes.add(attribute)
@@ -571,7 +572,7 @@ def test_sort_product_by_attribute_using_attribute_having_no_products(
     """Ensure passing an empty attribute ID as sorting field does nothing."""
 
     query = QUERY_SORT_PRODUCTS_BY_ATTRIBUTE
-    attribute_without_products = product_models.Attribute.objects.create(
+    attribute_without_products = attribute_models.Attribute.objects.create(
         name="Colors 2", slug="colors-2", type=AttributeType.PRODUCT_TYPE
     )
