+++ saleor/graphql/product/types/products.py
@@ -7,6 +7,7 @@ from graphene import relay
 from graphene_federation import key
 from graphql.error import GraphQLError
 
+from ....attribute import models as attribute_models
 from ....core.permissions import OrderPermissions, ProductPermissions
 from ....core.weight import convert_weight_to_default_weight_unit
 from ....product import models
@@ -102,7 +103,7 @@ def resolve_attribute_list(
         attributes_qs = attributes_qs.get_visible_to_user(user)
 
     # An empty QuerySet for unresolved values
-    empty_qs = models.AttributeValue.objects.none()
+    empty_qs = attribute_models.AttributeValue.objects.none()
 
     # Goes through all the attributes assigned to the product type
     # The assigned values are returned as a QuerySet, but will assign a
@@ -726,7 +727,9 @@ class ProductType(CountableDjangoObjectType):
     @staticmethod
     @permission_required(ProductPermissions.MANAGE_PRODUCTS)
     def resolve_available_attributes(root: models.ProductType, info, **kwargs):
-        qs = models.Attribute.objects.get_unassigned_product_type_attributes(root.pk)
+        qs = attribute_models.Attribute.objects.get_unassigned_product_type_attributes(
+            root.pk
+        )
         return resolve_attributes(info, qs=qs, **kwargs)
 
     @staticmethod
